+++ apps/extension/src/index.tsx
@@ -5,7 +5,7 @@ import { createRoot } from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import Global from './components/Global';
 import PopupRoutes from './components/PopupRoutes';
-import Logging from './error';
+import Logging from './logging';
 import DynamicProvider from './provider/DynamicProvider';
 
 Logging.logErrors();
