+++ apps/extension/src/logging.ts
